server/requirements.txt
gunicorn
django-heroku
python-dotenv
django-allauth
coverage
model-bakery
django-extensions
djangorestframework-simplejwt
coreapi
pyyaml
pyopenssl
social-auth-app-django
djongo
pytz